refactor(shared): extract toDateInputValue to shared/lib/time
Four modules independently re-implemented the same helper for
normalizing ISO date strings, plain date strings, and Date instances
into the YYYY-MM-DD shape that <input type="date"> requires:
- customers/DealForm
- staff/LeaveRequestForm
- staff/JobHistorySection
- sales/dashboard/shared (intentionally different — uses local
  timezone for dashboard range filtering, NOT touched here)

Move the non-dashboard variant to packages/shared/lib/time.ts alongside
formatDateTime and formatRelativeTime, add focused tests, and replace
the three local copies with shared imports. Call sites that expected
an empty string instead of null get a `?? ''` nullish coalescing.

Follow-up PRs will add the same import to workflows/formConfig.tsx and
business_rules/formHelpers.ts once the fixes in #1269 and #1273 merge.
