+++ packages/core/src/modules/staff/components/LeaveRequestForm.tsx
@@ -10,6 +10,7 @@ import {
   type UnavailabilityReasonEntry,
 } from '@open-mercato/core/modules/planner/components/unavailabilityReasons'
 import { useT } from '@open-mercato/shared/lib/i18n/context'
+import { toDateInputValue } from '@open-mercato/shared/lib/time'
 import { apiCall } from '@open-mercato/ui/backend/utils/apiCall'
 import { useOrganizationScopeVersion } from '@open-mercato/shared/lib/frontend/useOrganizationScope'
 
@@ -50,23 +51,6 @@ export type LeaveRequestFormProps = {
 }
 
 const DEFAULT_TIMEZONE = 'UTC'
-
-function toDateInputValue(value?: string | Date | null): string | null {
-  if (!value) return null
-  if (value instanceof Date) {
-    if (Number.isNaN(value.getTime())) return null
-    return value.toISOString().slice(0, 10)
-  }
-  if (typeof value === 'string') {
-    const trimmed = value.trim()
-    if (!trimmed) return null
-    if (/^\d{4}-\d{2}-\d{2}/.test(trimmed)) return trimmed.slice(0, 10)
-    const parsed = new Date(trimmed)
-    if (Number.isNaN(parsed.getTime())) return null
-    return parsed.toISOString().slice(0, 10)
-  }
-  return null
-}
 
 export function buildLeaveRequestPayload(
   values: LeaveRequestFormValues,
