+++ packages/shared/src/lib/__tests__/time.test.ts
@@ -1,4 +1,4 @@
-import { formatRelativeTime, formatDateTime } from '../time'
+import { formatRelativeTime, formatDateTime, toDateInputValue } from '../time'
 
 describe('formatRelativeTime', () => {
   const now = new Date('2026-02-18T12:00:00Z')
@@ -108,5 +108,30 @@ describe('formatDateTime', () => {
   it('returns a locale string for valid date', () => {
     const d = new Date('2026-02-18T12:00:00Z')
     expect(formatDateTime(d.toISOString())).toEqual(d.toLocaleString())
+  })
+})
+
+describe('toDateInputValue', () => {
+  it('returns null for missing or invalid values', () => {
+    expect(toDateInputValue(undefined)).toBeNull()
+    expect(toDateInputValue(null)).toBeNull()
+    expect(toDateInputValue('')).toBeNull()
+    expect(toDateInputValue('   ')).toBeNull()
+    expect(toDateInputValue('not-a-date')).toBeNull()
+    expect(toDateInputValue(new Date('not-a-date'))).toBeNull()
+  })
+
+  it('passes through a plain YYYY-MM-DD string unchanged', () => {
+    expect(toDateInputValue('2026-04-10')).toBe('2026-04-10')
+  })
+
+  it('slices the date portion of an ISO 8601 string', () => {
+    expect(toDateInputValue('2026-04-10T23:00:00.000Z')).toBe('2026-04-10')
+    expect(toDateInputValue('2026-04-10T00:00:00+02:00')).toBe('2026-04-10')
+  })
+
+  it('formats a Date instance as the UTC calendar day', () => {
+    const date = new Date('2026-04-10T00:00:00.000Z')
+    expect(toDateInputValue(date)).toBe('2026-04-10')
   })
 })
